Move P2P network initialization into network files
